actual changes, windows testing now works with gdb
