+++ libaudioviz/deps.cmake
@@ -3,7 +3,7 @@ include(FetchContent)
 
 ## GLEW (caused by pbo optimization changes)
 if(WIN32)
-	message("windows detected... fetching glew binaries")
+	message(STATUS "windows detected... fetching glew binaries")
 	FetchContent_Declare(glew URL https://github.com/nigels-com/glew/releases/download/glew-2.3.0/glew-2.3.0-win32.zip)
 	FetchContent_MakeAvailable(glew)
 	target_include_directories(audioviz PUBLIC ${glew_SOURCE_DIR}/include)
@@ -36,10 +36,6 @@ if(WIN32 AND NOT FFTW3_FOUND)
 		target_link_directories(audioviz PUBLIC ${fftw_SOURCE_DIR})
 		target_include_directories(audioviz PUBLIC ${fftw_SOURCE_DIR})
 		target_link_libraries(audioviz PUBLIC fftw3f-3)
-
-		# this is so that we only need to append the build dir to the PATH before running,
-		# making that process a bit less error-prone
-		file(COPY ${fftw_SOURCE_DIR}/libfftw3f-3.dll DESTINATION ${CMAKE_CURRENT_BINARY_DIR})
 	else()
 		message("other architecture detected, fetching fftw source...")
 		set(BUILD_TESTS OFF)
@@ -55,14 +51,23 @@ else()
 endif()
 
 ## sfml (without audio and network libs)
+set(SFML_STATIC_LIBRARIES ON)
 find_package(SFML COMPONENTS Graphics Window System QUIET)
 if(NOT SFML_FOUND)
-	message(STATUS "SFML not found, using FetchContent...")
-    set(SFML_BUILD_AUDIO OFF)
-    set(SFML_BUILD_NETWORK OFF)
-	FetchContent_Declare(SFML URL https://github.com/SFML/SFML/archive/3.0.2.tar.gz)
-	FetchContent_MakeAvailable(SFML)
+	if(WIN32)
+		message(STATUS "windows detected, downloading SFML binaries from GitHub Releases")
+		FetchContent_Declare(SFML URL https://github.com/SFML/SFML/releases/download/3.0.2/SFML-3.0.2-Windows.MinGW.x64.zip)
+		FetchContent_MakeAvailable(SFML)
+		list(APPEND CMAKE_MODULE_PATH ${SFML_SOURCE_DIR})
+	else()
+		message(STATUS "SFML not found, using FetchContent...")
+		set(SFML_BUILD_AUDIO OFF)
+		set(SFML_BUILD_NETWORK OFF)
+		FetchContent_Declare(SFML URL https://github.com/SFML/SFML/archive/3.0.2.tar.gz)
+		FetchContent_MakeAvailable(SFML)
+	endif()
 endif()
+find_package(SFML COMPONENTS Graphics Window System REQUIRED)
 target_link_libraries(audioviz PUBLIC SFML::Graphics)
 
 ## spline
