+++ libaudioviz/include/audioviz/Sprite.hpp
@@ -9,6 +9,7 @@ class Sprite : public sf::Sprite
 {
 public:
 	Sprite(const sf::Texture &txr);
+
 	// for the sole purpose of widescreen youtube thumbnails containing square album covers
 	void capture_centered_square_view();
 	void scale_to(const sf::Vector2f size);
