+++ libaudioviz/src/audioviz/fx/Add.cpp
@@ -18,7 +18,7 @@ void Add::apply(RenderTexture &rt) const
 {
 	shader.setUniform("size", sf::Glsl::Vec2{rt.getSize()});
 	shader.setUniform("addend", addend);
-	rt.draw(rt.sprite(), &shader);
+	rt.draw(rt, &shader);
 	rt.display();
 }
